An issue fixed, and new issues identified
POLARIS does not have a C+17 compiler, only a C+11 compiler. Any version of LAMMPS that needs the -DLAMMPS_CXX11=yes flag to compile on a C+11 compiler WILL NOT work on POLARIS.
Meanwhile, in the Elastic_WLC folder, performing a conjugate-descent minimization with stationary endpoints and arclengthrestraint fix will not cause the end-to-end arc length to converge to our desired total length. Measuring the forces on the atoms immediately after the minimization ends shows that the forces are not consistent with what was intended of the arclengthrestraint fix.
Performing damped dynamics causes the forces on the atoms to explode
(and so do their positions, unless one uses a very large value of the
damping constant).
This suggests that I need a serious reevaluation of the
arclengthrestraint fix.