Finally fix_arc_length_restraint works as intended
Because the configuration manifold that minimizes the restraint energy
is flat, you cannot use sd or hftn minimizer styles. If you use the cg
minimizer style, you must use the backtrack linesearch setting.

While the end-to-end arc length of the molecule post-minimization is
very close (within 0.3%) of our desired value, the end-to-end arc length
post-minimization is not evenly distributed amongst the bonds of the
molecule. This can lead to surprisingly-large forces on one of the atoms
that makes the post-minimization configuration not yet suitable for
MD runs.